Avoid unbalanced html, insert colorful markers.
Instead of the ins/del html elements insert raquo/laquo symbols.
Rationale: the diff logic does not know about well-formed html.
Thus you end up with a file that contains the diff information but
that cannot be rendered correctly by most html viewers. Using
independent symbols to mark start and end of some modification
reduces the danger of broken html. Granted, the modified parts are
not enclosed by ins/del tags anymore but the human eye is quite
capable to recognize start and end of the modified parts when they
are marked with colored arrow-like symbols.